Fix to separate classes
